requirements_pgo.txt
# Requirements for OpenGS-SLAM with PGO (Pose Graph Optimization)

# Core dependencies (from original environment)
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0
tqdm>=4.60.0
PyYAML>=6.0
wandb>=0.12.0

# PGO-specific dependencies
ultralytics>=8.0.0  # YOLOv8 for object detection
g2o-python>=0.1.0  # Pose graph optimization (may need to build from source)

# Optional: Alternative PGO libraries if g2o-python is not available
# ceres-python>=0.1.0  # Alternative optimization library
# gtsam>=4.0.0  # Another alternative for pose graph optimization

# Additional utilities
scikit-image>=0.19.0  # For image processing
scipy>=1.7.0  # For scientific computing
matplotlib>=3.5.0  # For visualization

# Development dependencies
pytest>=6.0.0
black>=22.0.0
flake8>=4.0.0
